Raise the maximum line length for checkstyle
This has not much effect, as the formatter will automatically format the lines and wrap them at 120.
However, in certain cases, the formatter can find in difficult to wrap the line (for example with a variable declaration and cast with a fully qualified name) and there is either a need to manually force the wrap (using an end of line comment for example) or disabling the check with a @SuppressWarning( "checkstyle:LineLength" ) annotation.
This change only removes the checks so that in those rare cases, the formatter can be left without any need to force things.